requirements.txt
django==5.2.7 
djangorestframework==3.16.1 
requests==2.32.5 
django-environ==0.12.0
django-cors-headers==4.9.0

